tango_with_django_19-master/code/requirements.txt
-f /usr/share/pip-wheels
Django==1.9.6
django-bootstrap-toolkit==2.15.0
django-registration-redux==1.4
Pillow==3.3.0
